+++ src/sys/os/browser/web/api/namespace.rs
@@ -10,7 +10,7 @@ use crate::_js_extern;
 crate::items! {
     use crate::{WebDocument, WebPermission, WebPermissionState, WebWindow};
     use crate::{_js_doc, js_int32};
-    #[cfg(feature = "event")]
+    #[cfg(all(feature = "event", feature = "time"))]
     use crate::{WebEventKind, js_uint32};
     #[cfg(feature = "time")]
     use crate::{JsInstant, js_number};